v5

+++ src/main/java/md/program/modelFX/PartnerModel.java
@@ -1,6 +1,8 @@
 package md.program.modelFX;
 
+import md.program.database.model.Partner;
 import md.program.database.repository.PartnerRepository;
+import md.program.database.repository.PaymentPlanRepository;
 import md.program.utils.converters.PartnerConverter;
 
 import java.sql.SQLException;
@@ -9,6 +11,7 @@ public class PartnerModel {
 
     private PartnerFX partnerFX = new PartnerFX();
     private PartnerRepository partnerRepository = new PartnerRepository();
+    private PaymentPlanRepository paymentPlanRepository = new PaymentPlanRepository();
 
     public PartnerFX getPartnerFX() {
         return partnerFX;
@@ -21,12 +24,29 @@ public class PartnerModel {
     public void setId() throws SQLException {
         partnerFX.setId(partnerRepository.getNextId());
     }
+
+    public Boolean validData() {
+        Partner partner = PartnerConverter.convertToPartner(partnerFX);
+        if (partner.getSurname().isEmpty() || partner.getAddress().isEmpty() || partner.getPostCode().isEmpty() || partner.getPost().isEmpty()) return false;
+        if (partner.getCompany() == false && partner.getPeopleCount() == 0) return false;
+        if (partner.getCompany()== true && partner.getNip().isEmpty()) return false;
+        return true;
+    }
+
     public void addPartner() throws SQLException {
-        partnerRepository.addPartner(PartnerConverter.convertToPartner(partnerFX));
+        Partner partner = PartnerConverter.convertToPartner(partnerFX);
+        if (partner.getCompany()) partner.setPeopleCount(0);
+        partnerRepository.addPartner(partner);
     }
-    public void deletePartner() throws SQLException {
-        partnerRepository.deleteRateYearById(PartnerConverter.convertToPartner(partnerFX));
+
+    public Boolean deletePartner() throws SQLException {
+        Integer temp = paymentPlanRepository.checkPartnerInPaymentPlan(PartnerConverter.convertToPartner(partnerFX));
+        if (temp == 0) {
+            partnerRepository.deleteRateYearById(PartnerConverter.convertToPartner(partnerFX));
+            return true;
+        } else return false;
     }
+
     public void saveUpdatePartner() throws SQLException {
         partnerRepository.updatePartner(PartnerConverter.convertToPartner(partnerFX));
     }
